requirements.txt
# Interfaz gráfica
PyQt5>=5.15.0

# Base de datos
mysql-connector-python>=8.0.0

# Utilidades
python-dotenv>=1.0.0

# Generación de datos ficticios
Faker>=20.0.0

# Herramientas de desarrollo (opcionales)
pytest>=7.0.0
black>=24.3.0

# Nota: Los siguientes módulos son parte de la biblioteca estándar de Python
# y no necesitan instalación: threading, time, datetime, re, os, sys, logging
